Fix NFA combining methods to invalidate passed NFA
NFA methods `concat`,`union`,`kleene` take one or more NFA as an
argument. Usage of the object after it was passed to these methods is
disallowed and may result in unexpected behvior and side effects.

Added `_invalidate` method to mark a NFA object invalid.
`concat`,`union`,`kleene` methods invalidate their arguments, `accept`
method checks for validity of the object it is invoked on. If invoked on
invalid object it raises `NFAInvalidError`.

If user intends to pass a NFA object to invalidating methods and use it
afterwards, one should pass a deep copy. Using `deepcopy` from `copy`
package is sufficient.

Closes #10