Add character set parsing and regex generation
Added positive character set support along with character ranges inside
a set. The implementation creates a union of all characters in a set,
which results in a large number of edges, but does not require any
changes to the NFA implementation.

Limitations, quirks and features of positive character sets:
* character set parser does not check for duplicate characters and
  happily creates a union with duplicates
  `[aabca-c]` -> `abc`
* empty character set is allowed and no character can match it
  `[]`
* character ranges in decreasing order result in zero characters
  `[z-a]` -> `[]`
* character ranges might yield one character
  `[a-a]` -> `a`
* character set parser does not check for unsupported or special
  characters
  `[*+.<-?]` -> `*+.<=>?`
* probably some other subtle issues

Negative character set support was not implemented, and parser raises
`RegexParseError` when one is encountered. There is no obvious way to
implement this feature without changes to the NFA implementation.

This concludes the initial regex parser implementation. Additional
features and enhancements might be added in future.

Closes #15
Closes #2