Add test for deep copying of NFA objects
NFA combining methods invalidate passed arguments. A deep copy is
needed if one wants to use an object after it was passed to one of
these methods.

Added a test which checks if internal State objects of the original and
the copy are different objects. The test also checks if both NFAs still
work as intended.

With 8a4f616 closes #13