Improve character literal parsing
Supported character literals used to be alphanumberic only. This commit
adds support for symbols and escaped special characters.

Symbol characters which don't have special meaning are added to the
supported character list `supported_chars` in regex.py and can be used
literally in a regex pattern. Special characters can be used as literals
if escaped with backslash.

NFA objects' descriptions are modifed to display an escaped form of a
literal.

Range expressions are not affected by this change.

Closes #28