Fix empty regex parsing bug
Empty regex was parsed deep in the parsing code. After all alternatives
was tried, and the las character was empty it was assumed that we ended
that way because whole pattern was empty, and returning `('empty',)` AST
was correct response.
It was incorrect assumption. The code for empty pattern was executed in
other patterns, for example `a|b|` which isn't allowed according to the
grammar.
Empty pattern handling is now the first thing before the actual parsing.

Closes #36