Fix unmatched parentheses bug
An unmatched right parenthesis caused the parser to accept an invalid
regex pattern and ignore the rest of the pattern after it.

Right parenthesis could only be consumed if there was a left parenthesis
earlier (grouping, see `elementaryRE`). It could also interrupt parsing
if it was the first character in a pattern, see `symbolRE` where it
would raise `RegexParseError` due to incorrect use of special character.

Otherwise an unmatched right parenthesis would end up interrupting the
parsing in `concatRE`, because concatenation must end if encountered "|"
or ")".

The fix is straightforward: check if `last_char` is ")" after the parser
finishes. Successful parsing should end with `last_char` empty.

Closes #39