Add `concat_many` NFA combination method
`concat_many` method returns a NFA which is equivalent to chaining
`concat` of all NFAs in the `concat_list` parameter.
The method requires all input NFAs to be valid, and doesn't invalidate
or use any NFAs if this requirement is not fulfilled.

If used on a list with only one NFA, `concat-many` returns the same NFA,
which is consistent with concatenation semantics.

Empty NFA list is not allowed.

Part of #47 and #30