Add `union_many` NFA combination method
`union_many` method returns a NFA which accepts a union of every NFA
passed in the `union_list`. Internally it uses a better technique than
chaining `union` operation over the list. It uses only two states total,
and two edges per passed NFA.

The method requires all input NFAs to be valid, and doesn't invalidate
or use any NFAs if this requirement is not fulfilled.

If used on a list with only one NFA, `union_many` returns an equivalent
NFA, which is consistent with union semantics.

Empty NFA list results in `none` NFA.

Part of #48 and #30