Add `char_set` NFA generation method
`char_set` method returns a NFA which accepts a character set regular
expression. Internally it uses a `union_many` NFA, added in 2d11c6a.

`char_set` takes two arguments:
* `char_list`, a list of character literals
* `description`, a description for returned NFA

`description` is assumed to be provided from the outside, because only
the parser knows an accurate description for a character set. With
current parser `char_set` could only give a list of all characters as a
description, because this is its entire knowledge.

Empty NFA list results in `none` NFA, a correct behavior for an empty
set.

Part of #49 and #30