Add DFA-based string matching support to Regex
Issue #18 and #19 allowed NFA-to-DFA conversion. DFA-based matchers can
be used in the `Regex` class for string matching.

Added `_matcher_type` attribute to the `Regex`, which describes the
current matcher.

Added `compile_to_DFA` method, which converts the current NFA matcher to
a DFA one. The DFA matcher is used in subsequent calls to `accept`.

Closes #59