Add usable registers information
`compile_to_x86` was added with information which registers should it
use to store variables, and which are to be preserved across a function
call. Compiled code might need to use register's lowest byte, but ESI
and EDI do not allow it in 32bit mode, therefore are excluded from
available registers list.

Part of #72