Add reg-mem (base+index) addressing test
Test if a move from a memory location specified by a base register and
an index register is encoded correctly. Encoding should use a modR/M and
SIB bytes. There is a special case for EBP base register without
displacement - it has to be encoded with a 1-byte displacement (with 0
value).

Part of #78