Add reg-mem (base only) addressing test
Test if a move from a memory location specified by a base register is
encoded correctly. Encoding should use a modR/M byte. There is special
case for ESP base register - it has to be encoded with a SIB byte. There
is also a special case for EBP base register without displacement - it
has to be encoded with a 1-byte displacement (with 0 value).

Part of #78