Change ESP/R12 indexing handling to exception
`Encoder.encode_instruction` was using `assert` to check that ESP/R12 is
never used as an index in memory addressing, because instruction
encoding in x86 doesn't allow it. This is changed to raising an
exception, because this can happen due to external code, e.g. bad code
generator tries to use this register as an index.

Part of #92