Add dummy encoder classes for future changes
Currently `Encoder` uses composition of methods instead of good old
inheritance. This change adds dummy classes `Encoder32`, `Encoder64`
which should be used instead of calling `Encoder('32')` and
`Encoder('64')` respectively.

These classes don't do more than calling `Encoder` with the correct
parameter, but this will be changed to use proper inheritance in future,
which should be an invisible change for outer code.

Part of #92