Move register extracting from arch-specific code
Extracting register code from a variable is needed on x86_64
architecture and it is a noop on x86_32, so this can be safetly moved
outside of architecture specific code.

Whole register extracting is needed, because third bit (counting from
zero) is used to distingiush R8-R15 registers, which are encoded with
the same codes as RAX,RCX,...RSI registers. Register code extracting leaves
only three bits, which are actual register codes.

Part of #92