Add wrapper for encoding x86_64 `inc` instruction
This change is a part of the new instruction encoding interface. Calling
`Encoder.encode_instruction` directly will be discouraged in favour of
wrapper methods like this.

x86_64 doesn't support 0x40+register `inc` opcode, which is now used as
REX prefix. This requires separate `enc_inc` wrappers for `Encoder32`
and `Encoder64`.

Part of #92