Add `imm_size` param to `encode_instruction`
New param is added because in x86 immediate value is not always the same
size as other operands. `size` param is the size of operands and
`imm_size` is the size of immediate value. `imm_size` defaults to `size`
if not present, so no old code should break.

Example:
cmp r/m16/32 imm8
cmp eax 127 -> 83 F8 7F
cmp ax 127 -> 66 83 F8 7F

Part of #92